+++ src/main/java/me/itzsomebody/radon/transformers/obfuscators/BadAnnotation.java
@@ -50,6 +50,7 @@ public class BadAnnotation extends Transformer {
 
     @Override
     public void setConfiguration(Map<String, Object> config) {
+        // Not needed
     }
 
     @Override
